Define separate build targets for some backends
Not all of the backends can be used freely. While libdispatch is
usually a permissively licensed platform library, Qt5 is not and
rather requires xdispatch to follow LGPL terms as Qt5::Core is
always linked.

If a project wants to use one build of xdispatch2 but not link against
Qt5 for all of the projects it had to do two builds of xdispatch2
before.

Now it is possible to select by either linking the xdispatch target or
the xdispatch_qt5 target respectively.