naive: Remove use of yield
Using yield to give CPU time to another thread is an often used pattern
which is not ideal though because a) this is a syscall so potentially
expensive and b) on an SMP system this may just cause unecessary
swapping (and with that cache invalidation) of tasks for no good reason
as the other thread might just be active on a different CPU at the very
same time (causing yield not to help at all).

Use relaxed read access instead.