naive: Fix race between threads expiring and spawning
When the threadpool has been executing a bunch of work and then seen
no work just long enough for threads to start expiring it could happen
that the number of threads was suboptimal when work got scheduled at
that same time.

This was caused by threads on their way to expiring while work was
scheduled in parallel. The scheduler would then deem the threads as
still active and not spawn any new ones causing threads to be missing
and processing to continue in non ideal quantities.