naive: Optimize wakeup of semaphore waiters
If the semaphore was released faster than waiters were acquiring the
counter it could happen that only a single waiter was woken up even
though the count would allow for more wakeups than that.

Introduce a variable tracking the number of waiters so that if there
is a waiter it will be reliably woken up on a release.